first round of tests
failing case for the mentioned vulnerability, where you can
use a public rsa key and get validated via hmac

failing case for verifying with the private key, why does 
this not work?

note: this actually includes a fix. before this it was not
possible to use the RSA method for signing keys because
it was always defaulting to hmac